* master.kernel.org:/home/rmk/linux-2.6-arm:
  [ARM] arch/arm/kernel/dma-isa.c: named initializers
  [ARM] 3527/1: MPCore Boot Lockup Fix
  [ARM] arch/arm/kernel/process.c: Fix warning
  [ARM] 3526/1: ioremap should use vunmap instead of vfree on ARM
  [ARM] 3524/1: ARM EABI: more 64-bit aligned stack fixes
  [ARM] 3517/1: move definition of PROC_INFO_SZ from procinfo.h to asm-offsets.h

+++ arch/arm/kernel/process.c
@@ -311,7 +311,7 @@ void free_thread_info(struct thread_info *thread)
 		struct thread_info_list *th = &get_cpu_var(thread_info_list);
 		if (th->nr < EXTRA_TASK_STRUCT) {
 			unsigned long *p = (unsigned long *)thread;
-			p[0] = th->head;
+			p[0] = (unsigned long)th->head;
 			th->head = p;
 			th->nr += 1;
 			put_cpu_var(thread_info_list);
